requirements.txt
apache-beam[gcp]==2.49.0
pandas==2.0.3
SQLAlchemy==2.0.19
oracledb==1.3.2
gcsfs==2023.6.0
google-cloud-bigquery==3.11.3
google-cloud-storage==2.10.0
google-cloud-secret-manager==2.16.2